* split unit tests for schema from dj/ray_dataset
